+++ test/roller/prodrr.sql
@@ -1,8 +1,8 @@
--- MySQL dump 10.13  Distrib 5.7.22, for osx10.12 (x86_64)
+-- MySQL dump 10.13  Distrib 5.7.22, for Linux (x86_64)
 --
 -- Host: localhost    Database: rentroll
 -- ------------------------------------------------------
--- Server version	5.7.22
+-- Server version	5.7.22-0ubuntu0.16.04.1-log
 
 /*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
 /*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
@@ -2403,7 +2403,7 @@ CREATE TABLE `Transactant` (
   `LastName` varchar(100) NOT NULL DEFAULT '',
   `PreferredName` varchar(100) NOT NULL DEFAULT '',
   `CompanyName` varchar(100) NOT NULL DEFAULT '',
-  `IsCompany` smallint(6) NOT NULL DEFAULT '0',
+  `IsCompany` tinyint(1) NOT NULL DEFAULT '0',
   `PrimaryEmail` varchar(100) NOT NULL DEFAULT '',
   `SecondaryEmail` varchar(100) NOT NULL DEFAULT '',
   `WorkPhone` varchar(100) NOT NULL DEFAULT '',
@@ -2518,4 +2518,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2018-06-07 21:03:05
+-- Dump completed on 2018-06-13 18:40:01
